Fix Time.EpochMS
Regression from 1c3521b26a8a12e1652837222c1154705e9d29dc

+++ src/core/shards/time.cpp
@@ -64,7 +64,7 @@ struct EpochMs {
   SHVar activate(SHContext *context, const SHVar &input) {
     using namespace std::chrono;
     milliseconds ms = duration_cast<milliseconds>(system_clock::now().time_since_epoch());
-    return Var(int(ms.count()));
+    return Var(int64_t(ms.count()));
   }
 };
